agent-trader/requirements.txt
flask>=3.0.0
yfinance>=0.2.36
feedparser>=6.0.10
requests>=2.31.0
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
gunicorn>=21.2.0
